refactor admin orders page structure
