+++ docs/exec-plans/index.md
@@ -6,6 +6,7 @@ Active and completed ExecPlans live here. The structure and required sections ar
 
 ## Active
 
+- [Admin Orders Page Slimming](./active/admin-orders-page-slimming.md)
 - [Agent-First Harness Phase 2](./active/agent-first-harness-phase-2.md)
 - [ExecPlan: Billing Credit Notifications](./active/billing-credit-notifications.md)
 - [Creator Dashboard Course Ownership Optimization](./active/creator-dashboard-course-ownership-optimization.md)
